Refactor router structure for cleaner codebase
In this commit, the structure for FastAPI routers has been consolidated and simplified to better accommodate future development. Specifically, the previously separate account related router files (create.py, get.py, update.py, image.py and account.py) have been merged into a single file named account.py, with router becoming a top-level object. This refactoring makes the codebase much cleaner and easier to navigate. The same approach was applied for the product related routers. Changes were also made in entries.py to adapt to the new file structure. As for functionality, everything remains the same. This change was necessary for maintaining an organized codebase and allowing easier addition of new APIs in the future.
