--- router/Product/product.py
@@ -1,7 +1,0 @@
-from fastapi import APIRouter
-
-prefix = "/product"
-router = APIRouter(
-    prefix=prefix,
-    tags=["product"],
-)
